Create fields for notes.
Adds a private notes field to book and holding. Makes the notes
field private for transactions and inventories.

fixes sfu-dhil/bep#62
